move process_cases to last remaining place its used
