+++ corehq/apps/cleanup/management/commands/reprocess_error_forms.py
@@ -1,6 +1,6 @@
 from collections import defaultdict
 from django.core.management.base import BaseCommand, CommandError, LabelCommand
-from casexml.apps.case.util import reprocess_form_cases
+from corehq.apps.cleanup.management.commands.reprocess_error_form import reprocess_form_cases
 from corehq.apps.cleanup.xforms import iter_problem_forms
 from optparse import make_option
 from dimagi.utils.parsing import string_to_datetime
