+++ corehq/ex-submodules/casexml/apps/case/util.py
@@ -76,26 +76,6 @@ def create_real_cases_from_dummy_cases(cases):
     return posted_forms, posted_cases
 
 
-def reprocess_form_cases(form, config=None, case_db=None):
-    """
-    For a given form, reprocess all case elements inside it. This operation
-    should be a no-op if the form was sucessfully processed, but should
-    correctly inject the update into the case history if the form was NOT
-    successfully processed.
-    """
-    from casexml.apps.case.xform import process_cases, process_cases_with_casedb
-
-    if case_db:
-        process_cases_with_casedb([form], case_db, config=config)
-    else:
-        process_cases(form, config)
-    # mark cleaned up now that we've reprocessed it
-    if form.doc_type != 'XFormInstance':
-        form = XFormInstance.get(form._id)
-        form.doc_type = 'XFormInstance'
-        form.save()
-
-
 def get_case_xform_ids(case_id):
     results = XFormInstance.get_db().view('case/form_case_index',
                                           reduce=False,
@@ -124,13 +104,14 @@ def update_sync_log_with_checks(sync_log, xform, cases, case_db,
                 if form_id != xform._id:
                     form = XFormInstance.get(form_id)
                     if form.doc_type == 'XFormInstance':
-                        reprocess_form_cases(
-                            form,
+                        from casexml.apps.case.xform import process_cases_with_casedb
+                        process_cases_with_casedb(
+                            [form],
+                            case_db,
                             CaseProcessingConfig(
                                 strict_asserts=True,
                                 case_id_blacklist=case_id_blacklist
-                            ),
-                            case_db=case_db
+                            )
                         )
             updated_log = get_properly_wrapped_sync_log(sync_log._id)
